Transaction: Map `evmc_result` back to `Computation` result
This innocuous lookup change allows third-party EVMs to work.
(As long as there are no nested calls or creates; `*CALL` or `CREATE*` ops).

From here on, we can use [`evmone`](https://github.com/ethereum/evmone) and get
useful results with tracing enabled (`show_tx_calls = true`).  There are issues
with account balances, etc that need to be worked on to get the correct
results, but this is a big step.

Signed-off-by: Jamie Lokier <jamie@shareable.org>