EVMC: Byte-endian conversions for 256-bit numbers on EVMC host side
Perform byte-endian conversion for 256-bit numeric values, but not 256-bit
hashes.  These conversions are necessary for EVMC binary compatibility.

In new EVMC, all conversions are explicit, calling `flip256`.

It is arguable whether conversion should occur for storage slots (`key`).  In
the EVMC definition these are `evmc_bytes32`, meaning endian conversion is not
expected.  But Nimbus currently treats these as numbers not blobs.  The
difference shows up when running the tests: The tests pass with byte-endian
conversion for storage slots.

These conversions are performed in the EVMC "glue" code, which deals with the
binary interface, so the host services aren't aware of conversions.

The plan is to skip these conversions when Nimbus host is calling Nimbus EVM,
even when the EVM is dynamically loaded.  It will use a negotiated extension.
But for now we're focused on compatibility with third party EVMs.

The effective overhead of endian conversion is not too high, just because most
EVMC host calls access the database anyway.  `getTxContext` does not, but well
optimised EVMs don't call it repeatedly.

Signed-off-by: Jamie Lokier <jamie@shareable.org>