Sync: Robust support for `GetNodeData` network calls
This module provides an async function to call `GetNodeData`, a request in
the Ethereum DevP2P/ETH network protocol.  Parallel requests can be issued,
maintaining a pipeline.

Given a list of hashes, it returns a list of trie nodes or contract
bytecodes matching those hashes.  The returned nodes may be any subset of
those requested, including an empty list.  The returned nodes are not
necessarily in the same order as the request, so a mapping from request
items to node items is included.  On timeout or error, `nil` is returned.

Only data passing hash verification is returned, so hashes don't need to be
verified again.  No exceptions are raised, and no detail is returned about
timeouts or errors, but systematically formatted trace messages are output
if enabled, and show in detail if various events occur such as timeouts,
bad hashes, mixed replies, network errors, etc.

This tracks queued requests and individual request hashes, verifies received
node data hashes, and matches them against requests.  When a peer replies in
same order as requests are sent, and each reply contains nodes in the same
order as requested, the matching process is efficient.  It avoids storing
request details in a hash table when possible.  If replies or nodes are out
of order, the process is still efficient but has to do a little more work.

Empty replies:

Empty replies are matched with requests using a queue draining technique.
After an empty reply is received, we temporarily pause further requests and
wait for more replies.  After we receive all outstanding replies, we know
which requests the empty replies were for, and can complete those requests.

Eth/66 protocol:

Although Eth/66 simplifies by matching replies to requests, replies can still
have data out of order or missing, so hashes still need to be verified and
looked up.  Much of the code here is still required for Eth/66.

References:

- [Ethereum Wire Protocol (ETH)](https://github.com/ethereum/devp2p/blob/master/caps/eth.md)
- [`GetNodeData` (0x0d)](https://github.com/ethereum/devp2p/blob/master/caps/eth.md#getnodedata-0x0d)
- [`NodeData` (0x0e)](https://github.com/ethereum/devp2p/blob/master/caps/eth.md#nodedata-0x0e)