Sync: Robustly parse trie nodes from network untrusted data
This module parses hexary trie nodes as used by Ethereum from data received
over the network.  The data is untrusted, and a non-canonical RLP encoding
of the node must be rejected, so it is parsed carefully.

The caller provides bytes and context.  Context includes node hash, trie
path, and a boolean saying if this trie node is child of an extension node.

The result of parsing is up to 16 child node hashes to follow, or up to 7
leaf nodes to decode.

The caller should ensure the bytes are verified against the hash before
calling this parser.  Even though they pass the hash, they are still
untrusted bytes that must be parsed carefully, because the hash itself is
from an untrusted source.

`RlpError` exceptions may occur on some well-crafted adversarial input
due to the RLP reader implementation.  They could be trapped and treated
like other parse errors, but they are not, to avoid the overhead of
`try..except` in the parser (which uses C `setjmp`).  The caller should
put `try..except RlpError` outside its trie node parsing loop.